Update code post review
