--- examples/with_openai/with_openai/partitions.py
@@ -1,5 +1,0 @@
-from dagster import StaticPartitionsDefinition
-
-docs_partitions_def = StaticPartitionsDefinition(
-    ["concepts", "dagster-cloud", "deployment", "guides", "integrations"]
-)
