--- examples/with_openai/with_openai/resources.py
@@ -1,4 +1,0 @@
-from dagster import EnvVar
-from dagster_openai import OpenAIResource
-
-openai_resource = OpenAIResource(api_key=EnvVar("OPENAI_API_KEY"))
